+++ types/type_name.h
@@ -1,14 +1,18 @@
 //* type_name
 //      this header provides a type name function which creates a constexpr string view with the name of a type
-//      type_name_n returns the name of the type including all namespaces
-//      type_name returns the name of the type excluding fresa namespaces
-//      relevant discussion in https://stackoverflow.com/questions/81870/is-it-possible-to-print-a-variables-type-in-standard-c
+//      it also includes functionality to create type hashes
 #pragma once
 
 #include "std_types.h"
+#include "strong_types.h"
 
 namespace fresa
 {
+    //* type name
+    //      type_name_n returns the name of the type including all namespaces
+    //      type_name returns the name of the type excluding fresa namespaces
+    //      relevant discussion in https://stackoverflow.com/questions/81870/is-it-possible-to-print-a-variables-type-in-standard-c
+
     //: forward declaration of type_name_t to allow overloading
     template <typename T> constexpr str_view type_name_n();
 
@@ -61,5 +65,26 @@ namespace fresa
             return name.substr(pos + 1);
         }
         return name;
+    }
+
+    //* type hash
+    //      hashes the type name to allow for type lookup in an unordered map
+
+    constexpr ui64 hash_fnv1a(str_view s) {
+        constexpr ui64 fnv_prime = 0x100000001b3;
+        ui64 result = 0xcbf29ce484222325;
+        for (auto c : s) {
+            result ^= c;
+            result *= fnv_prime;
+        }
+        return result;
+    }
+
+    using TypeHash = strong::Type<ui64, decltype([]{}), strong::Hashable>;
+
+    template <typename T>
+    [[nodiscard]] constexpr TypeHash type_hash() {
+        constexpr auto name = type_name_n<T>();
+        return hash_fnv1a(name);
     }
 }